+++ src/test/java/de/alpharogroup/bean/mapper/MapperExtensionsTest.java
@@ -24,43 +24,32 @@
  */
 package de.alpharogroup.bean.mapper;
 
-import java.util.ArrayList;
-import java.util.List;
+import static org.testng.Assert.assertEquals;
 
-import org.dozer.DozerBeanMapper;
+import org.testng.annotations.Test;
 
 import de.alpharogroup.bean.mapper.factories.MapperFactory;
-import lombok.NonNull;
+import de.alpharogroup.test.objects.Member;
+import de.alpharogroup.test.objects.Person;
+import de.alpharogroup.test.objects.enums.Gender;
 
-/**
- * The class {@link DozerBeanMapperSingleton} holds a single instance of the class
- * {@linkplain DozerBeanMapper}.
- */
-public final class DozerBeanMapperSingleton
+public class MapperExtensionsTest
 {
 
-	/**
-	 * The only single one instance.
-	 */
-	private static DozerBeanMapper instance;
-
-	public static synchronized DozerBeanMapper getInstance(final @NonNull Class<?> entityClass,
-		final @NonNull Class<?> dtoClass)
+	@Test
+	public void testMap()
 	{
-		if (instance == null)
-		{
-			List<String> mappingFiles = new ArrayList<>();
-			mappingFiles.add("uuid-mapping.xml");
-			instance = (DozerBeanMapper)MapperFactory.newMapper(mappingFiles, entityClass,
-				dtoClass);
-		}
-		return instance;
 	}
 
-	/**
-	 * Instantiates a new {@link DozerBeanMapperSingleton}.
-	 */
-	private DozerBeanMapperSingleton()
+	@Test
+	public void testTestMap()
 	{
+		Member actual;
+		Member expected;
+		Person asterix = Person.builder().name("asterix").build();
+		actual = MapperExtensions.map(MapperFactory.newMapper(), asterix, Member.class);
+		expected = Member.buildMember().about("").name("asterix").gender(Gender.UNDEFINED)
+			.married(false).nickname("").build();
+		assertEquals(actual, expected);
 	}
 }